fix: handle form validation when form is initialised via the redo action flow
