+++ src/components/v5/common/ActionSidebar/partials/forms/ManagePermissionsForm/utils.ts
@@ -1,10 +1,14 @@
 import { ColonyRole, Id } from '@colony/colony-js';
+import { type UseFormSetValue } from 'react-hook-form';
 
 import {
   CUSTOM_USER_ROLE,
   UserRole,
   USER_ROLES,
+  getRole,
 } from '~constants/permissions.ts';
+import { type ColonyFragment } from '~gql';
+import { getUserRolesForDomain } from '~transformers';
 import { type Colony } from '~types/graphql.ts';
 import { getEnumValueFromKey } from '~utils/getEnumValueFromKey.ts';
 import { formatText } from '~utils/intl.ts';
@@ -107,4 +111,49 @@ export const extractColonyRoleFromPermissionKey = (permissionKey: string) => {
   console.error('Manage Permissions Form: Invalid permission: ', permissionKey);
 
   return null;
+};
+
+export const configureFormRoles = ({
+  colony,
+  setValue,
+  isSubmitted,
+  member,
+  role,
+  team,
+}: {
+  colony: ColonyFragment;
+  setValue: UseFormSetValue<ManagePermissionsFormValues>;
+  isSubmitted: boolean;
+  member: ManagePermissionsFormValues['member'];
+  team: ManagePermissionsFormValues['team'];
+  role: ManagePermissionsFormValues['role'];
+  shouldPersistRole?: boolean;
+  setShouldPersistRole?: React.Dispatch<React.SetStateAction<boolean>>;
+}) => {
+  const userRolesForDomain = getUserRolesForDomain({
+    colony,
+    userAddress: member,
+    domainId: team,
+    intersectingRoles: true,
+  });
+
+  const userRoleMeta = getRole(userRolesForDomain);
+
+  const userRole = userRoleMeta.permissions.length
+    ? userRoleMeta.role
+    : undefined;
+
+  setValue('_dbUserRole', userRole);
+  setValue('_dbUserPermissions', userRolesForDomain);
+
+  if (role !== UserRole.Custom) {
+    setValue('role', userRole, { shouldValidate: isSubmitted });
+  }
+
+  AVAILABLE_ROLES.forEach((colonyRole) => {
+    setValue(
+      `permissions.role_${colonyRole}`,
+      userRoleMeta.permissions.includes(colonyRole),
+    );
+  });
 };
